# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "gem",
        "description": "gem is project about transportation",
        "version": "1"
    },
    "paths": {
...
        "/api/drivers": {
            "get": {
                "tags": [
                    "drivers"
                ],
                "summary": "get all drivers",
                "description": "get all drivers",
                "operationId": "a75c8627a59b2479de780c37b0c06e03",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "drivers"
                ],
                "summary": "create driver",
                "description": "create driver",
                "operationId": "829ed7a657b54c9727b94e7b814c2497",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "panel_code": {
                                        "type": "integer"
                                    },
                                    "perosnal_code": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "father_name": {
                                        "type": "integer"
                                    },
                                    "national_code": {
                                        "type": "string"
                                    },
                                    "birth_certificate_code": {
                                        "type": "string"
                                    },
                                    "driver_licence_number": {
                                        "type": "string"
                                    },
                                    "driver_licence_expire": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "driver_licence_type": {
                                        "type": "integer"
                                    },
                                    "health_card_number": {
                                        "type": "string"
                                    },
                                    "health_card_expire": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "insurance_number": {
                                        "type": "string"
                                    },
                                    "smart_number": {
                                        "type": "string"
                                    },
                                    "smart_number_expire": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "vehicle_smart_number": {
                                        "type": "integer"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "city_of_birhth": {
                                        "type": "string"
                                    },
                                    "city_of_driver_licence": {
                                        "type": "string"
                                    },
                                    "allowed_vehicle_type": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "active": {
                                        "type": "boolean"
                                    },
                                    "country_code": {
                                        "type": "integer"
                                    },
                                    "country": {
                                        "type": "string"
                                    },
                                    "organization_phone": {
                                        "type": "string"
                                    },
                                    "phones": {
                                        "type": "string"
                                    },
                                    "postal_code": {
                                        "type": "string"
                                    },
                                    "start_activity": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "province": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "education": {
                                        "type": "string"
                                    },
                                    "IBAN": {
                                        "type": "string"
                                    },
                                    "last_inquiry_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Success accept",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/drivers/{driver}": {
            "get": {
                "tags": [
                    "drivers"
                ],
                "summary": "get driver by id",
                "description": "get driver by id",
                "operationId": "932db1d376b0d2c39ea954063cc0b390",
                "parameters": [
                    {
                        "name": "driver",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "drivers"
                ],
                "summary": "delete driver by id",
                "description": "delete driver by id",
                "operationId": "ac7a4a6a3d6bdd4145338527b23f76da",
                "parameters": [
                    {
                        "name": "driver",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success (without content)",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "drivers"
                ],
                "summary": "update driver",
                "description": "update driver",
                "operationId": "98875cef74527687fe108a6dc6715cad",
                "parameters": [
                    {
                        "name": "driver",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "panel_code": {
                                        "type": "integer"
                                    },
                                    "perosnal_code": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "father_name": {
                                        "type": "integer"
                                    },
                                    "national_code": {
                                        "type": "string"
                                    },
                                    "birth_certificate_code": {
                                        "type": "string"
                                    },
                                    "driver_licence_number": {
                                        "type": "string"
                                    },
                                    "driver_licence_expire": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "driver_licence_type": {
                                        "type": "integer"
                                    },
                                    "health_card_number": {
                                        "type": "string"
                                    },
                                    "health_card_expire": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "insurance_number": {
                                        "type": "string"
                                    },
                                    "smart_number": {
                                        "type": "string"
                                    },
                                    "smart_number_expire": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "vehicle_smart_number": {
                                        "type": "integer"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "city_of_birhth": {
                                        "type": "string"
                                    },
                                    "city_of_driver_licence": {
                                        "type": "string"
                                    },
                                    "allowed_vehicle_type": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "active": {
                                        "type": "boolean"
                                    },
                                    "country_code": {
                                        "type": "integer"
                                    },
                                    "country": {
                                        "type": "string"
                                    },
                                    "organization_phone": {
                                        "type": "string"
                                    },
                                    "phones": {
                                        "type": "string"
                                    },
                                    "postal_code": {
                                        "type": "string"
                                    },
                                    "start_activity": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "province": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "education": {
                                        "type": "string"
                                    },
                                    "IBAN": {
                                        "type": "string"
                                    },
                                    "last_inquiry_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Success accept",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/drivers/load-driver-by-national-code": {
            "post": {
                "tags": [
                    "drivers"
                ],
                "summary": "loading driver by national code",
                "description": "loading driver by national code",
                "operationId": "ec44439ebc12819538e2273bb9354d8c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "national_code": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    }
}